requirements.txt
beautifulsoup4==4.11.1
Django==3.2.16
djangorestframework==3.13.1
djangorestframework_simplejwt==5.3.1
drf_yasg==1.21.7
folium==0.15.1
geopy==2.4.1
googletrans==4.0.0rc1
multi_rake==0.0.2
nltk==3.8.1
numpy==1.24.1
pandas==1.5.2
ratelimiter==1.2.0.post0
Requests==2.32.3
scipy==1.14.0
tqdm==4.64.1
